requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx 
pytest
python-dotenv